extracted duplicate encrypted asset deletion to a function
