+++ build.gradle.kts
@@ -1,5 +1,4 @@
-import java.io.FileInputStream
-import java.io.FileOutputStream
+import Util.cleanEncryptedAssets
 
 buildscript {
     repositories {
@@ -23,26 +22,6 @@ allprojects {
 
 tasks.register<Delete>("clean") {
     delete(rootProject.buildDir)
-
-    val tempAssets = File(projectDir, "/src/main/assets-temp")
-    if (tempAssets.exists()) {
-        println("cleaning encrypted assets...")
-
-        File(projectDir, "src/main/assets").deleteRecursively()
-
-        tempAssets.walkTopDown().filter { it.isFile }.forEach { file ->
-            val fo = File(file.absolutePath.replace("assets-temp", "assets"))
-                .apply {
-                    parentFile.mkdirs()
-                }
-
-            FileInputStream(file).use { fis ->
-                FileOutputStream(fo).use { fos ->
-                    fis.copyTo(fos, bufferSize = 4096)
-                }
-            }
-        }
-        tempAssets.deleteRecursively()
-    }
+    cleanEncryptedAssets(projectDir)
 }
